Added new Preference dialog
This new dialog wil let the user set his/her preferred hotkey
combination. I was able to implement a fun way of setting it by simply
the desired keys instead of clicking via mouse which is pretty neat.
However, Tenny is crashing after showing this new dialog but this will
be fix soon.