Changed the keyComboBox to keyLineEdit to show the single_key
The original idea is to have a LineEdit that read the keys entered in the Set Shortcut dialog. But I don't know how to implement it back then. So decided to use a ComboBox and manually list all the available keys on the keyboard. Then I recently discovered that the keyboard library has a read_key() method that listens to the keys pressed. It was a relief to me when I discovered the method. That is the feature I want to implement while setting the shortcut for Tenny. Instead of manually selecting the list of available keys in a ComboBox, a LineEdit will now show the text representation of the key pressed. This approach saves a few lines of codes and execution time. There are still pending enhancements that I want to imlement but this will be the highlight for version 0.3. Tenny has now entered the beta testing phase.