Successfully making Tenny 'run' in the background
I thought I am going to imlement a thread to make the application run in the background, but after digging for answers in the net I was able to find the solution. The trick is to simply hide Tenny in the close event of the window. But before hiding the window, we need to know what triggered the application to close. Does the user simply pressed the close button or uses the quit menu in the notification area to exit the app? To achieve that, I use self.sender() and checked the type of the sender if it's a QAction. This means that the user wants to close the application. If not, just ignore the close event via ignore method. I think that's all for today. I learned another lesson. Coding is f*cking fun and somehow addicting!!