Fix QSS file when running live
Just like before, the app crashes when running standalone due to
FileNotFoundError. Fix this by adding a snippet if the app is running in
freeze mode or development and let PyInstaller to add 'qss/style.qss'
while while freezing.

Learned all of this using SO and a bunch of patience way back in
developing previous side projects. I'm so glad it help today.